refactor: :recycle: Create new Mazic Class
