+++ src/Character.py
@@ -1,4 +1,4 @@
-from src.Logger import Logger
+from Logger import Logger
 
 
 class Character:
